Workaround attempt in order to fix cf-mysql-deployment empty directory issue with circle CI

+++ cf-ops-automation-broker-core/src/main/java/com/orange/oss/cloudfoundry/broker/opsautomation/ondemandbroker/pipeline/tools/Copy.java
@@ -144,7 +144,9 @@ public class Copy {
             if (Files.isSymbolicLink(file)){
                 createSymbolicLink(file, target, root);
             }else {
-                copyFile(file, target.resolve(source.relativize(file)), prompt, preserve);
+                if (!file.getFileName().toString().contentEquals(".keep")){
+                    copyFile(file, target.resolve(source.relativize(file)), prompt, preserve);
+                }
             }
             return CONTINUE;
         }
